fix unassign prof

+++ src/app/departement/list-departements/list-departements.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
 import { Departement } from 'src/app/core/model/departement';
 import { Enseignant } from 'src/app/core/model/Enseignant';
 import { DepartementService } from 'src/app/core/services/departement.service';
@@ -15,24 +16,37 @@ export class ListDepartementsComponent implements OnInit {
 public list: Departement[];
 public listlength:number=0;
 searchtext:any;
-  constructor(private depService:DepartementService, private ensService:EnseignantService, private route: ActivatedRoute) { }
+  constructor(private depService:DepartementService, private ensService:EnseignantService, private route: ActivatedRoute,private toastr: ToastrService) { }
 
   ngOnInit(): void {
     this.depService.getAllDepartement().subscribe(
       (response:Departement[]) => { this.list = response ;
-      },
+      console.log(this.list), this.listlength=this.list.length},
       () => { console.log("error") },
       () => { console.log("complete") },
       
     )
   }
+
+  
+
+
+
+
+
 deleteDep(d:Departement){
   let i=this.list.indexOf(d);
+  if(confirm("Are you sure to delete "+d.nomDepart)) {
+
   this.depService.deleteDepartement(d.idDepartement).subscribe(
-    ()=>{this.list.splice(i,1), this.listlength=this.list.length},
-    ()=>{console.log("error while deleting department")}
-  )
-}
+    ()=>{this.list.splice(i,1), this.listlength=this.list.length,  this.toastr.success(d.nomDepart+' has been deleted successfully','Success');
+  },
+    error => (err: string) => {
+      console.log("err" + err);
+      this.toastr.error('something went wrong !', 'Error');
+    }
+  )}}
+
   getColor(departement: Departement) { 
   switch (departement.bloc) {
     case 'A':
